# src/main/resources/swagger.yml
...
swagger: '2.0'
info:
  description: Documentation for GymTracker API
  version: '1.0'
  title: GymTracker API Documentation
host: localhost:20183
basePath: "/"
...
paths:
  "/api/auth/login":
    post:
      tags:
        - Authorization API
      summary: Login to the application
      operationId: loginUsingPOST
      consumes:
        - application/json
      produces:
        - "*/*"
      parameters:
        - in: body
          name: userLoginDto
          description: userLoginDto
          required: true
          schema:
            "$ref": "#/definitions/UserLoginDto"
      responses:
        '200':
          description: OK
          schema:
            type: string
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  "/api/auth/register":
    post:
      tags:
        - Authorization API
      summary: Register a new user
      operationId: registerUsingPOST
      consumes:
        - application/json
      produces:
        - "*/*"
      parameters:
        - in: body
          name: userRegisterDto
          description: userRegisterDto
          required: true
          schema:
            "$ref": "#/definitions/UserRegisterDto"
      responses:
        '200':
          description: OK
          schema:
            "$ref": "#/definitions/SuccessResponse"
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  "/api/exercises":
    get:
      tags:
        - Exercise API
      summary: Get list of all exercises
      operationId: getExercisesUsingGET
      produces:
        - "*/*"
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              "$ref": "#/definitions/ExerciseResponseDto"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - Exercise API
      summary: Create a new exercise
      operationId: createExerciseUsingPOST
      consumes:
        - application/json
      produces:
        - "*/*"
      parameters:
        - in: body
          name: exerciseDto
          description: exerciseDto
          required: true
          schema:
            "$ref": "#/definitions/ExerciseDto"
      responses:
        '200':
          description: OK
          schema:
            "$ref": "#/definitions/SuccessResponse"
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  "/api/exercises/{id}":
    put:
      tags:
        - Exercise API
      summary: Edit an existing exercise
      operationId: editExerciseUsingPUT
      consumes:
        - application/json
      produces:
        - "*/*"
      parameters:
        - in: body
          name: exerciseDto
          description: exerciseDto
          required: true
          schema:
            "$ref": "#/definitions/ExerciseDto"
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            "$ref": "#/definitions/SuccessResponse"
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - Exercise API
      summary: Delete an exercise
      operationId: deleteExerciseUsingDELETE
      produces:
        - "*/*"
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            "$ref": "#/definitions/SuccessResponse"
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  "/api/tickets":
    get:
      tags:
        - Ticket API
      summary: Get all tickets
      operationId: getAllTicketsUsingGET
      produces:
        - "*/*"
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              "$ref": "#/definitions/TicketResponseDto"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - Ticket API
      summary: Create a new ticket
      operationId: createTicketUsingPOST
      consumes:
        - application/json
      produces:
        - "*/*"
      parameters:
        - in: body
          name: ticketDto
          description: ticketDto
          required: true
          schema:
            "$ref": "#/definitions/TicketDto"
      responses:
        '200':
          description: OK
          schema:
            "$ref": "#/definitions/SuccessResponse"
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  "/api/tickets/{id}":
    put:
      tags:
        - Ticket API
      summary: Edit a ticket
      operationId: editTicketUsingPUT
      consumes:
        - application/json
      produces:
        - "*/*"
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
        - in: body
          name: ticketDto
          description: ticketDto
          required: true
          schema:
            "$ref": "#/definitions/TicketDto"
      responses:
        '200':
          description: OK
          schema:
            "$ref": "#/definitions/SuccessResponse"
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - Ticket API
      summary: Delete a ticket
      operationId: deleteTicketUsingDELETE
      produces:
        - "*/*"
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            "$ref": "#/definitions/SuccessResponse"
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  "/api/training-logs/gym-diary/{trainingSessionId}":
    get:
      tags:
        - Training Log API
      summary: Get training logs for a training session
      operationId: getTrainingLogsForTrainingSessionUsingGET
      produces:
        - "*/*"
      parameters:
        - name: trainingSessionId
          in: path
          description: trainingSessionId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              "$ref": "#/definitions/TrainingLogResponseDto"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  "/api/training-logs/{id}":
    put:
      tags:
        - Training Log API
      summary: Edit a training log
      operationId: editTrainingLogUsingPUT
      consumes:
        - application/json
      produces:
        - "*/*"
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
        - in: body
          name: trainingLogDto
          description: trainingLogDto
          required: true
          schema:
            "$ref": "#/definitions/TrainingLogDto"
      responses:
        '200':
          description: OK
          schema:
            "$ref": "#/definitions/SuccessResponse"
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - Training Log API
      summary: Delete a training log
      operationId: deleteTrainingLogByIdUsingDELETE
      produces:
        - "*/*"
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            "$ref": "#/definitions/SuccessResponse"
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
definitions:
...
  ExerciseDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      muscleGroup:
        type: string
        enum:
          - ABS
          - BACK
          - BICEPS
          - CALVES
          - CHEST
          - GLUTES
          - LEGS
          - SHOULDERS
          - TRICEPS
      name:
        type: string
    title: ExerciseDto
  ExerciseResponseDto:
    type: object
    properties:
      muscleGroup:
        type: string
        enum:
          - ABS
          - BACK
          - BICEPS
          - CALVES
          - CHEST
          - GLUTES
          - LEGS
          - SHOULDERS
          - TRICEPS
      name:
        type: string
    title: ExerciseResponseDto
...
  SuccessResponse:
    type: object
    properties:
      body:
        type: object
      httpStatus:
        type: string
        enum:
          - ACCEPTED
          - ALREADY_REPORTED
          - BAD_GATEWAY
          - BAD_REQUEST
          - BANDWIDTH_LIMIT_EXCEEDED
          - CHECKPOINT
          - CONFLICT
          - CONTINUE
          - CREATED
          - DESTINATION_LOCKED
          - EXPECTATION_FAILED
          - FAILED_DEPENDENCY
          - FORBIDDEN
          - FOUND
          - GATEWAY_TIMEOUT
          - GONE
          - HTTP_VERSION_NOT_SUPPORTED
          - IM_USED
          - INSUFFICIENT_SPACE_ON_RESOURCE
          - INSUFFICIENT_STORAGE
          - INTERNAL_SERVER_ERROR
          - I_AM_A_TEAPOT
          - LENGTH_REQUIRED
          - LOCKED
          - LOOP_DETECTED
          - METHOD_FAILURE
          - METHOD_NOT_ALLOWED
          - MOVED_PERMANENTLY
          - MOVED_TEMPORARILY
          - MULTIPLE_CHOICES
          - MULTI_STATUS
          - NETWORK_AUTHENTICATION_REQUIRED
          - NON_AUTHORITATIVE_INFORMATION
          - NOT_ACCEPTABLE
          - NOT_EXTENDED
          - NOT_FOUND
          - NOT_IMPLEMENTED
          - NOT_MODIFIED
          - NO_CONTENT
          - OK
          - PARTIAL_CONTENT
          - PAYLOAD_TOO_LARGE
          - PAYMENT_REQUIRED
          - PERMANENT_REDIRECT
          - PRECONDITION_FAILED
          - PRECONDITION_REQUIRED
          - PROCESSING
          - PROXY_AUTHENTICATION_REQUIRED
          - REQUESTED_RANGE_NOT_SATISFIABLE
          - REQUEST_ENTITY_TOO_LARGE
          - REQUEST_HEADER_FIELDS_TOO_LARGE
          - REQUEST_TIMEOUT
          - REQUEST_URI_TOO_LONG
          - RESET_CONTENT
          - SEE_OTHER
          - SERVICE_UNAVAILABLE
          - SWITCHING_PROTOCOLS
          - TEMPORARY_REDIRECT
          - TOO_EARLY
          - TOO_MANY_REQUESTS
          - UNAUTHORIZED
          - UNAVAILABLE_FOR_LEGAL_REASONS
          - UNPROCESSABLE_ENTITY
          - UNSUPPORTED_MEDIA_TYPE
          - UPGRADE_REQUIRED
          - URI_TOO_LONG
          - USE_PROXY
          - VARIANT_ALSO_NEGOTIATES
      timeStamp:
        type: string
        format: date-time
    title: SuccessResponse
...
  TicketDto:
    type: object
    required:
      - description
      - subject
    properties:
      description:
        type: string
      subject:
        type: string
    title: TicketDto
  TicketResponseDto:
    type: object
    properties:
      creationDate:
        type: string
        format: date-time
      description:
        type: string
      subject:
        type: string
    title: TicketResponseDto
...
  TrainingLogDto:
    type: object
    required:
      - exerciseId
    properties:
      exerciseId:
        type: integer
        format: int64
      personalNotes:
        type: string
      reps:
        type: integer
        format: int32
      trainingSessionId:
        type: integer
        format: int64
      weight:
        type: integer
        format: int32
    title: TrainingLogDto
  TrainingLogResponseDto:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      personalNotes:
        type: string
      reps:
        type: integer
        format: int64
      weight:
        type: integer
        format: int64
    title: TrainingLogResponseDto
...
  UserLoginDto:
    type: object
    properties:
      login:
        type: string
      password:
        type: string
    title: UserLoginDto
  UserRegisterDto:
    type: object
    required:
      - emailAddress
      - login
      - password
    properties:
      emailAddress:
        type: string
      login:
        type: string
      password:
        type: string
    title: UserRegisterDto
...